Seperate Dasboard items into different components.
